requirements.txt
# requirements.txt - minimal & reproducible set for experiments
# Python 3.10 is recommended.

# core ML / DL
torch>=2.1.0
torchvision>=0.16.0

# data science & processing
numpy>=1.24
pandas>=2.0
scipy>=1.10
scikit-learn>=1.2
scikit-image>=0.20
joblib

# image & augmentation
pillow>=9.0
opencv-python>=4.6
albumentations>=1.3

# explainability & metrics
captum>=0.6
lpips>=0.1.4  # perceptual similarity (optional, used for reconstruction metrics)

# utils & logging
tqdm
tensorboard
pyyaml
wandb  # optional; remove if you don't want to use Weights & Biases

# plotting & viz
matplotlib>=3.5
seaborn

# dev / testing
pytest  # optional: for unit tests
black   # optional: code formatting
